Add cp_model patch fixtures and simplify optimiser tests
